+++ spark/src/test/scala/org/radiant/opendatalake/normalized/GnomadCNVV1Spec.scala
@@ -39,8 +39,8 @@ class GnomadCNVV1Spec extends SparkSpec {
     del.getAs[String]("name") shouldBe "variant_is_80_2__DEL"
     del.getAs[String]("svtype") shouldBe "DEL"
     del.getAs[Int]("svlen") shouldBe 5553
-    del.getAs[Double]("sc") shouldBe 3.0
-    del.getAs[Double]("sn") shouldBe 464277.0
+    del.getAs[Long]("sc") shouldBe 3L
+    del.getAs[Long]("sn") shouldBe 464277L
     del.getAs[Double]("sf") shouldBe 6.46165974192131e-06
   }
 
@@ -48,8 +48,8 @@ class GnomadCNVV1Spec extends SparkSpec {
     val fixture = getClass.getResource("/input_vcf/gnomadV4CNV.vcf").getPath
     val raw = spark.read.format("vcf").option("flattenInfoFields", "true").load(fixture)
 
-    // The release carries about 112 INFO fields; publishing one commits the contract to it, since
-    // dropping a column later is a MAJOR while adding one is a MINOR carried by schema evolution.
+    // The source VCF declares 100 INFO fields. We verify that only the 11 fields mentioned in the contract
+    // are published.
     val fields = job.transformSingle(Map(source.id -> raw)).schema.fieldNames
     fields should have length 11
     fields.filter(f => f.contains("_xx") || f.contains("_xy") || f.contains("_nfe")) shouldBe empty
@@ -63,8 +63,6 @@ class GnomadCNVV1Spec extends SparkSpec {
     val fixture = getClass.getResource("/input_vcf/gnomadV4CNV.vcf").getPath
     val raw = spark.read.format("vcf").option("flattenInfoFields", "true").load(fixture)
 
-    // FILTER holds PASS or FAIL only — 747 of the 68,699 released sites are FAIL, masked for a noisy
-    // normalized read-depth signal. They are dropped, which is why `filters` is not published.
     val failed = raw.withColumn("filters", array(lit("FAIL")))
 
     job.transformSingle(Map(source.id -> failed)).count() shouldBe 0
